+++ UI/GameEditor/ViewModels/ModulesViewModel.cs
@@ -1,11 +1,10 @@
 ﻿namespace Macabre2D.UI.GameEditor.ViewModels {
 
     using Macabre2D.Framework;
-    using Macabre2D.UI.CommonLibrary.Models.FrameworkWrappers;
     using Macabre2D.UI.CommonLibrary.Services;
 
     public sealed class ModulesViewModel : NotifyPropertyChanged {
-        private ModuleWrapper _selectedModule;
+        private BaseModule _selectedModule;
 
         public ModulesViewModel(IBusyService busyService, ISceneService sceneService) {
             this.BusyService = busyService;
@@ -16,7 +15,7 @@
 
         public ISceneService SceneService { get; }
 
-        public ModuleWrapper SelectedModule {
+        public BaseModule SelectedModule {
             get {
                 return this._selectedModule;
             }